Add defining attributes to Person model
Feature:

When I look at the response plan for a person,
I want to see their defining characteristics
So that I can get a sense for what the person looks like.

Solution:

Add "horsepower" attributes to the `Person` model.

"Horsepower" is jargon that the Police use
to refer to peoples' defining characteristics.
This includes:

- sex
- race
- height
- weight
- hair_color
- eye_color
- date_of_birth

Since `height` and `weight` are numerical values,
we need to define which units we're using
when we store the value in the database.
To clarify, we've named the database columns
`height_in_inches` and `weight_in_pounds`.