Remove dependency on data.csv file
Problem:

We were previously pulling data out of a CSV file
that was exported from the crisis response template responses.
Because that file is not included in our source code,
the tests on CircleCI are failing.

Solution:

Remove the `data.csv` file,
and create a class that stubs out the future interaction
with the RMS system.