Remove the route at /latest
Problem:

We have a route at `/latest` that is no longer being used.

Previously, it would redirect the user to the most recent crisis
response plan filed. This was the location that the Google Form
redirected users to after they submitted a new response plan.

Since we are no longer using Google Forms, we have no immediate need for
a `/latest` route.

Solution:

Since the code is unused and untested, we've removed it.

Minor change:

Improve the URL lookup by person's name