Add response strategies (steps in a response plan)
Closes #20

Feature:

When I'm responding to a crisis call
I want to see personalized information
  about how to help the person in crisis
So that I can get the best possible outcome for them.

Implementation:

Create a `ResponseStrategy` model,
which represents a single strategy that an officer can use
to help a person in crisis.

Each response strategy has a priority,
which is used to decide the order in which the response plans appear on
the page.

At the moment, response strategies cannot be reordered.
That feature should be implemented when we build out the form interface
for CRT officers.