Lay out information in sections, remove cruft
Problem:

The page that shows individual response plans
is cluttered with information from our build week prototype,
which isn't necessarily relevant to our project
based on what we learned in Seattle in February.

As we work to implement features based on our designs and user testing,
the unused information is getting in the way.
It is difficult to move sections around on the page
and modify the styles in an ordered way
when there's a lot of extra HTML and CSS
that can potentially cause conflicts and unexpected behavior.

Solution:

- Remove unused information
- Move all remaining information into sections,
  following the UI language created in the design mocks.
- Move the person's image to the left side of the page,
  and all other information on the right.
- Extract mixins for each of the sections
- Remove unused CSS rules, organize the used ones

Removed information:

- safety warnings
- questions/contact information
- footer
- veteran status
- license number
- address
- needs
  - food
  - housing
- past encounter descriptions
- footer
  - GO number
  - date
  - prepared by
  - approved by

For each of these, we've removed:

- html
- css
- relevant images