Add a Docker environment for hosting the app
Problem:

We cannot install dependencies directly on the production environment
because it is not connected to the internet.

Solution:

We can package up our entire development environment with Docker
to avoid manually installing dependencies one-by-one.

We're using the basic Rails Docker configuration suggested here:
https://docs.docker.com/compose/rails/

Instructions for installing `capybara-webkit`'s dependencies are here:
https://robots.thoughtbot.com/rails-on-docker