Add safety warning categories to db and view
Closes #19

Feature:

When I'm driving to a crisis call,
I want to see potential safety warnings about the subject
so I can be prepared to interact with the person.

Implementation:

Add columns in our database for different safety warnings:

- gun
- knife
- needles
- other_weapon
- police_assault
- spitting
- suicide_by_cop
- violence

Each of the columns is boolean, non-nil, and defaults to `false`.

- `true` means "We know this is a threat"
- `false` means "We have no reason to suspect it's a threat"

Decisions & Tradeoffs:

There's an argument for allowing `nil` values in the columns,
which would essentially give officers an "I don't know" option
when filling out the form.
This sounds like a good idea – there are many cases when an officer
does not know whether a person is armed or not,
and does not want to represent that in the system.

But from a technical perspective, it introduces a lot of complexity
to allow a `nil` value for boolean columns.
In a lot of situations, Rails treats `nil` values the same as `false`,
which means that we would be very likely to introduce a bug
where an officer said "I don't know", but we treat it as if they said
"this is not a safety concern".

Having only two options for the safety warnings
helps us be a bit more certain about the values in our database.
As mentioned above, the working definitions are:

- `true` means "We know this is a threat"
- `false` means "We have no reason to suspect it's a threat"