Add safety warnings to db and view
Closes #19

Feature:
--------

When I'm driving to a crisis call,
I want to see potential safety warnings about the subject
so I can be prepared to interact with the person.

Implementation:
---------------

Create a `SafetyWarning` model backed by a `safety_warnings` table.
Each safety warning needs to be associated with a person,
and is required to have a description.

Alternative #1: Booleans
------------------------

We initially attempted to implement this with boolean columns,
as essentially a "yes/no" option for whether something was a threat.

The columns corresponded to warnings about:

- gun
- knife
- needles
- other_weapon
- police_assault
- spitting
- suicide_by_cop
- violence

However, an officer expressed during a usability test
that it would be good to see free-form information
for the safety warnings.

Situations are too unpredictable for checkboxes.

Alternative #2: Strings
-----------------------

Add columns in our database for different safety warnings:

- gun
- knife
- needles
- other_weapon
- police_assault
- spitting
- suicide_by_cop
- violence

Each of the columns is a string.
A value of `nil` for a column
means that there is no known safety warning in that category.

This approach was much more flexible,
but would likely result in more maintenance work
as new categories came along.