Display the physical characteristics in shorthand
Problem:

There is a lot of unused space on the physical characteristics card,
and much of the information is self-identifying
without the need for labels.

Solution:

List out the main information in the shorthand notation
that the officers are used to seeing.
For example, a white male who is 5'10" and weighs 180 lbs
would be described as: `WM – 5'10" – 180 lb`.
This format is space-efficient
and fits in well with their current practices.

For hair and eye color, distinguish between them with icons.

Identifying scars and marks are displayed at the bottom.