Add fuzzy searching by name and date of birth
Closes #80
Closes #81

Feature:
--------

**As** an officer,
**When I** am looking for a person's crisis response plan
**I want to** search by their name and date of birth
**So I can** quickly pull up their information using information I know.

Implementation:
---------------

We're using the excellent [pg_search] gem ([blog])
to take advantage of the built-in search functionality in Postgres.

Using `pg_search`, we've defined a `#search` method on `Person`
that takes a query and returns records
with a matching first or last name.
The method also supports searching over both column,
which is important because there is a single search box
for both first and last names.

The gem has built-in support for "fuzzy" matching.
We're using two search strategies for this:
[dmetaphone] matches words that are phonetically similar to the query,
and [trigram] corrects for small typo mistakes.

To make these work,
we needed to activate two postgres extensions on our database.
We followed [installation instructions on the project's Wiki][install]
to install the extensions.

For searching by date of birth,
we're simply returning people whose date of birth
is within one year of the date given.

We're using the [chronic] gem to parse the date query.
The gem sensibly handles any date format someone is likely to enter,
which makes it perfect for a search flow.

[pg_search]: https://github.com/Casecommons/pg_search
[blog]: https://blog.pivotal.io/labs/labs/pg-search-how-i-learned-to-stop-worrying-and-love-postgresql-full-text-search

[dmetaphone]: https://github.com/Casecommons/pg_search#dmetaphone-double-metaphone-soundalike-search
[trigram]: https://github.com/Casecommons/pg_search#trigram-trigram-search

[install]: https://github.com/Casecommons/pg_search/wiki/Installing-Postgres-Contrib-Modules

[chronic]: https://github.com/mojombo/chronic

Decisions & Tradeoffs:
----------------------

There are several searching solutions,
such as [elasticsearch] and [SOLR],
which are much more powerful than Postgres' built-in search abilities.
However, they also require running another service alongside our app.
This makes the application significantly harder to maintain,
and introduces a whole new class of potential bugs that can occur
when the search index gets out of sync with the main database.
By keeping our search index entirely within our database,
we will hopefully avoid those bugs.

Another option that we considered
which also uses Postgres' built-in search abilities
is the [textacular] gem.
After experimenting with it in our app,
We discovered that it would not search for a single query
across multiple columns.
In our app, that meant that officers would not be able to search
for both a person's first and last name at the same time.
That's a necessary feature, so we rejected Textacular.

[elasticsearch]: http://elasticsearch.org/
[SOLR]: https://lucene.apache.org/solr/
[textacular]: http://textacular.github.io/textacular/