Add emergency contact information
Feature:

When I am interacting with a person in crisis,
I want to quickly get in touch with people who know the person
so they can help respond to the crisis.

Implementation:

- Add a `Contact` model
- Associate the contact model with a `Person`
- Replace the UI placeholders with the contact information

Small changes:

Add a `tel_to` helper in the `ApplicationHelper` module,
which takes a phone number and turns it into a clickable link.

See http://makandracards.com/makandra/29353-tel_to-rails-helper-for-linking-phone-numbers