Make app runnable both in Docker and locally
Problem:

Our application relied on libv8 as a javascript runtime in Docker
(as a dependency of the gem`therubyracer`),
but it wasn't readily installable on my local machine.
This means we needed to run all rake or rails commands
within Docker, which is a bit cumbersome for a dev workflow.

Solution:

* Install nodejs as a javascript runtime in the Docker image,
  so we don't need to rely on libv8.
* Remove the dependency on `therubyracer`,
  and let developers choose their own javascript runtimes.
* Because we can now run the tests in multiple environments,
  the user can now specify which database to use
  through the `DB_HOST` and `DB_USER` environment variables.

Notes:

On local machines, users can install any javascript runtime
that is supported by [ExecJS].
At this time, that includes:

* therubyracer - Google V8 embedded within Ruby
* therubyrhino - Mozilla Rhino embedded within JRuby
* Node.js
* Apple JavaScriptCore - Included with Mac OS X
* Microsoft Windows Script Host (JScript)

[ExecJS]: https://github.com/sstephenson/execjs