Remove incident history section
Problem:
--------

When we display real data in the response plans,
we can't display dummy data for the incident history.
Inaccurate data is worse than no data.

Solution:
---------

Remove the "Incident History" section of the response plan.
Since we will likely redesign the section before implementing it,
it is probably safe to remove the charting library (chartist) and code.