Add a rake task to import CRP info from CSV files
The task can be run with the command:

```bash
rake import[data/response_plans.csv]
```

We've also refactored the database seeds file
to use the same importer as the rake task.

For images, the importer expects them stored in a directory structure
that matches the structure of the `data.sample` folder.
The folder for a person's images is calculated
from their first and last names,
and the first image alphabetically is the one that gets imported.