Display author & approver in response plan
Feature:

When I'm interacting with a person in crisis
And I need more information about the person
I want to contact the officer who created their response plan
So I can talk to them about the person.

Implementation:

Create an `Officer` model
with a name, title, unit, and phone number.

Associate each response plan with two officers –
the plan's officer, and the officer who approved the plan.

Plans are not visible to patrol officers until they have been approved.

A plan's author can not approve the response plan

Add a box to the bottom of the response plan page,
with the author's name, unit, and phone number.
Also display the name of the officer who approved the response plan.

Minor changes:

- Update seeds file to conform to new validations.
- Update seeds file to raise an error quickly when seeds are invalid