Add background information to response plans
Problem:

The "Response Plan" section doesn't have enough space or flexibility
for officers to enter all the information about a person.
Some information is useful, but is not directly related
to how officers should respond to the person.

Solution:

Create a "Background Information" section,
where the CRT officers can enter free-form information about a person
that is not related to their response plans.