Rename Person -> ResponsePlan
Problem:

Currently, the Person model stores information
both about the person in crisis, and their response plan.
Although both names are correct to an extent,
the model is being treated more like a response plan than a person,
especially when it comes to the "author"ing and approval process.

Solution:

Rename `Person` -> `ResponsePlan`.

There were a couple options for renaming this in the database schema.

The standard way to do this would be to create a new migration
that renames the `people` table to `response_plans`,
and renames the matching associations accordingly.
This approach would keep the existing data in place.

However, since we don't yet have production data
that we need to migrate,
we can take the simpler approach of renaming all of the old migrations.

This makes the database history a bit cleaner,
and is a bit easier than writing a migration from scratch.

The caveat is that in order to jump in the git history
from before this commit to after it, or vice versa,
the database will need to be entirely dropped and recreated.