Restructure response plan layout based on MDT size
Problem:
--------

The three-column layout that we were using
doesn't display well on the MDT's rather small screen (1024 px).

A large problem was that there wasn't enough room to display
the physical characteristics box
and the officer safety box
next to each other, where they both took up 1/3 of the screen.

Solution:
---------

During a "devsign" meeting,
Meredith and I looked at the page layout at the MDT screen size,
and went through several iterations to settle on a new layout.

The new layout is two-column,
with the person's identifying information and photos in the left column,
and narrative and response plan information on the right.

We moved the physical characteristics and person's name/DOB
into the "images" box, which we renamed to "profile".

Small Changes:
--------------

- Improve information hierarchy of response plan information
- Make bold text bolder throughout the plan, for better contrast
- Make the Safety Warnings section header red, to stand out
- remove "DOB" label from header