Extract an "identity" partial for name and DOB
Problem:

The header and profile sections display the same information –
name and date of birth.

At the moment, they aren't formatted correctly,
and they are displaying hard-coded information instead of data pulled
from the response plan.

Solution:

Extract a partial so both places display the name and DOB consistently.

The header applies some additional formatting to the "identity" partial
inside of it, to make the spacing work.