Improve layout, formatting of "about" section
Problem:

Formatting:

There's not much space in the "about" section,
so when we display labels side-by-side with the values,
the text tends to run into itself.

Layout:

On large screens where we can display two columns,
we want to display the "about" section
directly below the person's image.
On smaller screens where we can only fit one column,
we want to display the "about" section at the bottom of the page.
There's no straightforward way to do this with just CSS,
without getting hacky.

Solution:

Improve the formatting of text in the "about" section.

Render the "about" section HTML on the page twice,
and use css to hide the instance
that is in the wrong place for the current screen size.