Send feedback alert emails thru Toby's script
To test that we could send emails from our CentOS box,
Toby provided us with a script that sends emails from the command line.
The script makes a connection to the SPD's `smtp2` server
and writes the messages to its telnet buffer.

At least, that's the best I can tell.

Somehow, that script is the only way I've found
to send email from the SPD network.
The built-in Rails ActiveMailer system
simply fails to deliver a message to my inbox,
but does not report any errors in the program or configuration.

To get it working for our program,
we've created an adapter class (`EmailService`) that interfaces
between Rails email objects (such as an instance of `FeedbackMailer`)
and the command-line interface for Toby's script.
We can pass any Rails email object to `EmailService.send(message)`,
and it will use Toby's script to send the message.

At the moment, we are using this to send our team notification emails.
In the future, this system can be used for password resets,
bulletin update alerts, etc.