Replace the hamburger menu icon with a search icon
Problem:

The hamburger menu icon linked to the search page.
That's unintuitive for officers,
especially if they are not familiar with hamburger menus.

Solution:

Replace the hamburger menu with a magnifying glass search icon.

After we changed the header icon,
the tests got confused that there were now two links labeled "search"
on the search page.
The testing engine did not know whether to click on the header link
or the (correct) "Search" button in the form
when performing a search by name or date of birth.

To fix this, we restricted the tests to only look
in the `.search` form element
when submitting a search query.