Remove all unused ActionMailer SMTP configuration
Problem:

We're using an unconventional method of sending email,
based on a bash script that Toby wrote.

Since we're not using the built-in Rails email framework,
the configuration files for it are misleading to developers
who might come to the project from a standard Rails background.

Solution:

Remove the unused configurations from the app.