Add Keen.IO for analytics
Right now we're tracking page views, by:

- Officer ID
- Response plan ID
- Controller name
- Action name
- KEEN_ENV environment variable,
  to distinguish development from production
- HTML referrer

The build was failing for a while
because Poltergeist encounters a JS error
when it loads the keen-js library.

To fix this, we've split up the `application.js` file
into three separate, application-specific files.
We load Keen.io in the `development` and `production` environments,
but not in `test`.