:bug: make HTML field text visible in night mode
Fixes #147.

Problem:

In night mode, the default text color is set to white.
When it's rendered on the white background of form fields,
it is not visible to users.

Solution:

Explicitly set the color of form field text
in both night mode and day mode
so that it overrides the default text color.