Fix display problems on IE9
Problem:

The SPD's IE9 browsers are set to display any page on the intranet
in compatibility mode.

This means that the desktop browsers were displaying our page
as if it was being viewed on IE7.

Solution:

Add a meta tag specifying the compatibile version of IE
that the browser should use.
(See http://stackoverflow.com/questions/6348959)

Also, adjust the spacing of some of the header elements.