Generate analytics tokens for officers and plans
## Problem:

We need to track officers and response plans in our analytics,
but the officers' and subjects' identities are sensitive information.
We cannot directly share their names or usernames,
so thus far we've been using their database IDs.

Tracking based on database IDs is unsustainable,
because many routine database operations can re-assign those IDs.
When that happens, there would be a break in the analytics,
and we would not be able to use previously-collected analytics
with analytics collected after the change.

## Solution:

Generate random "analytics tokens" for response plans and officers,
which we can use to identify the records in our analytics.
The new analytics tokens can be carried over with the records
through any database changes,
and they don't contain any identifying information about officers
or subjects of the response plans.

We considered some other identifiers
as alternatives to the random tokens:

- name
- hashed name
- officer id
- username
- hashed username

The name and username are identifiable, so they can't be used.

The hashed name and username are not identifiable,
but they also cannot be easily queried in the database
because they are calculated values that aren't stored.

The officer ID is a four-digit alphanumeric string that appears before
the user's `displayName` in their LDAP entry.
This would be a promising candidate, but not all officers have an ID.