Load Keen.js from a CDN
## Problem:

When we loaded the Keen.io javascript from a rails-assets gem,
it broke our tests.
This forced us to separate our JS manifest files
so we would only require that library in development and production.

## Solution:

The Keen.io [installation instructions][gh] offer a CDN-hosted script
that is equivalent to the `rails-assets`-hosted gem.
We can include a call to that script that in our layout file
and remove the gem dependency.

[gh]: https://github.com/keen/keen-js/#install-the-library