Make officers sign in if their account is removed
## Problem:

If we remove an officer's account,
it will be re-created the next time an officer signs in.
The officer will have a new database id,
but the session's `officer_id` that identifies them
will remain as their old one.

This will almost certainly break the app,
as pages that expect to work with the officer database record
will instead be working with a `nil` object.

## Solution:

Only authenticate the officer if their ID matches a database record.