Add "about" and "changelog" pages
We parse markdown partials using the solution suggested [here].

The "About" page displays field tester information
that we can update as needed.

The "Change Log" page shows all of the changes to the app,
as described in the `CHANGELOG.md` file.

The pages are available at `/pages/about` and `/pages/changelog`.

To keep the Changelog page up to date with the project's CHANGELOG,
we've created a symlink to the CHANGELOG.md at the root of the repo
from the `app/pages` directory.
Any changes made to the CHANGELOG are always displayed in the app
with no additional work on our part.

[here]: http://stackoverflow.com/questions/4163560/10131299#10131299