Gracefully report malformed date of birth searches
## Problem:

If an officer searches for a date of birth in an unrecognized format,
the application throws a server error.

## Solution:

For malformed DOBs in searches,
ignore them when finding search results
and display an error to the user in the UI.

## Clean up

Remove some unnecessary factory attributes
that weren't relevant to the specs.