Don't run the specs through Docker
## Problem:

The phantomjs / poltergeist instance that we're using in Docker
throws errors when we ask it to click on links in the sliding menu bar.

Since our `bin/merge` script runs the tests through Docker,
this error gets in the way of the merging workflow.

## Solution:

The tests consistently pass outside of Docker,
so it's easiest to simply always run them outside.