Display locations with address and a map
## Feature

Officers often want to know where a person lives
or is frequently seen at.

## Implementation

Use the Google Maps API to render a nice-looking form card.
Made heavy use of the Google Maps JS library [documentation].

[documentation]: https://developers.google.com/maps/documentation/javascript/
[autocomplete]: https://developers.google.com/maps/documentation/javascript/places-autocomplete#address_forms

In the database, locations are stored as two columns
on the `response_plans` table – `location_name` and `location_address`.
To keep addresses as flexible as possible for the moment,
we're storing them as a single text field.
We let the Google Geocoder API do the hard work of normalizing it into a
standard address with lat & lng.