Clean up small details of UI
* Create night versions of the zoom in and out icons
* Add a night version of the phone icon
* Add organization to contacts
* Fix bugs with switching images in IE9

# IE9 image problems

## Problem:

IE9 does not support `dataset` and `classList` in its javascript.
These functions are both used for switching the focused image
on repsonse plans, so users are unable to switch images in IE9.

## Solution:

Use alternate methods for switching the images,
using solutions suggested on StackOverflow.

See:

http://stackoverflow.com/questions/8098406/code-with-classlist-does-not-work-in-ie#8098430
http://stackoverflow.com/questions/7127896/problem-using-elem-dataset-with-ie-and-jsfiddle

# Oganizations for contacts

This matches the mocks, but we are not yet importing the organizations
from the CSV file. We need to reformat the CSV to account for
organizations.