On small screens, collapse response plan steps
## Problem

On small screens, the text of response plan steps
takes up a lot of vertical space,
forcing the user to scroll far down the page
to get to following information.

## Solution

Display the first two lines of the response plan step descriptions,
with a link to toggle showing the full description.

## Implementation Notes

The designs called for ellipses for text that overflowed
the height of the collapsed container.
This is not possible to do with pure CSS,
and we would need to introduce a new dependency such as [ftellipsis].
More details are available on [StackOverflow][so].

[ftellipsis]: https://github.com/ftlabs/ftellipsis
[so]: http://stackoverflow.com/questions/7004006/is-vertical-text-overflow-possible-with-css3

Inspiration for the code to swap out the "expand" link text
came from [this article][swap].

[swap]: https://css-tricks.com/swapping-out-text-five-different-ways/

I attempted to animate height according to [this article][animate],
but it ended up being pretty jumpy and not smooth on mobile.
Decided against it.

[animate]: http://css3.bradshawenterprises.com/animating_height/