Switch between themes without a page refresh
## Problem:

It takes a long time to switch between night and day themes,
because it requires another request to the server.

## Solution:

When the user clicks the button to change themes,
immediately change the URL on the stylesheet link tag
to the other theme's stylesheet URL.

Send an AJAX post request to the server,
so it can record the user's preference.

See:
http://api.rubyonrails.org/classes/ActionView/Helpers/AssetUrlHelper.html#method-i-path_to_stylesheet