Transform the header with the menu bar
The mocks call for it.

Ran into a bit of a tricky situation with IE9 -
the header bar wouldn't translate with the rest of the content
in the `main-content` div.

This caused the header action links (search and menu)
to be obscured by the side menu.
Without being able to click on the menu button again,
users were not able to re-hide the menu.

It turns out this is because the header has `position: fixed`,
which is necessary to keep it visible at the top of the screen
as users scroll down.

IE prioritizes the `position: fixed` higher
than it prioritizes the transformation of the parent element.
So when the rest of the `.main-content` shifts left,
the header is left in its fixed position.

We get around this by explicitly shifting the header as well.
We apply the `-ms-` vendor prefix to that transformation
to avoid double-shifting the header too far to the left.

See:
https://developer.mozilla.org/en-US/docs/Web/CSS/transform?redirectlocale=en-US&redirectslug=CSS%2Ftransform