Fix up test suite poltergeist warnings
## Problem:

The Google Maps Javascript API prints warning messages
when the element it is being applied to
does not appear on the screen.

Inside Docker, these warnings cause the test suite to fail.

## Solution:

Check that the relevant element is present on the page
before attempting to draw a map on it.