Adjust Docker configuration to push to Docker Hub
The image is `codeforamerica/crisisresponse`.

https://hub.docker.com/r/codeforamerica/crisisresponse/

## Problem

It takes a long time to move our application's Docker image
from the development machine to the host server.

Right now we are packaging the image into a `.tar` archive file,
placing it on a flash drive,
and then using WinSCP to copy the file onto the server
through a remote desktop connection.

Anytime a change is made to the Docker image,
even for a relatively small update like a new Gemfile dependency,
the entire 1gb Docker image must be re-compressed and re-transferred.
This makes for some very slow deployments.

## Solution

Push the Docker image to Docker Hub.

Docker Hub caches the different layers of the Docker image,
so the consumer only needs to pull the changed layers
instead of the entire Docker image.

As long as our server has a connection to the internet,
this drastically speeds up our deployments.