Extract a `Person` model from `ResponsePlan`
## Problem

When a response plan is updated,
we need to be able to show the previous version of the plan
while the new version is pending approval.

We need to have a place to store information from the RMS,
in a way that can be easily updated by a regular ETL script
without affecting the data displayed to officers.

## Solution

We should have two tables,
`Person` and `RMSPersonRecord`,
which both store information about individuals
the police have interacted with.

The `RMSPersonRecord` table will hold information
directly pulled from the RMS,
while the `Person` table will hold custom information
entered by the Crisis Response Team.

The `Person` table will have the ability to be associated with
an `RMSPersonRecord`.
If it is associated, we can use the information stored in the
`RMSPersonRecord` to suggest or autocomplete
information about the person in our app.

The first step along this path is extracting a `Person` model out of
`ResponsePlan`, to hold the person's physical characteristics and other
information that doesn't change between versions of a response plan.

This will also let us track versions of response plan information,
which will help us comply with public disclosure requests and the legal
auditing requirements of the Police Department.