Automatically fill dashes for Date of Birth field
## Problem

When officers fill in a date of birth,
they need to enter nubers as well as punctuation ('-', '/').
Switching between the number and symbol keyboards
is time-consuming and annoying.

## Solution

Automatically insert the appropriate punctuation for the date,
along with a placeholder template that shows the expected date format.

As officers fill out the field,
they can enter only numbers
and their input will fall into the correct format.

We're using the [JQuery Masked Input Plugin][plugin]
to handle the details of masking the field.

[plugin]: https://github.com/digitalBush/jquery.maskedinput