Use rails-assets to install jquery, jquery-ujs
## Problem

The `jquery.maskedinput` library from `rails-assets.org`
depends on `jquery`,
and installs a version of the library from `rails-assets.org`.
This overlaps the version of jquery installed by Rails
in Ruby's `jquery-rails` gem.

It is unusual, and likely a bad idea to have two versions of `jquery`
installed side-by-side on a single project.

## Solution

Replace the default `jquery` installed by Rails
with the version from `rails-assets.org`.

The version from `rails-assets.org` (3.1.0) is newer
than the version installed by `jquery-rails` (version 2).

To keep everything working happily,
we also needed to change `jquery-ujs` to come from `rails-assets.org`.

There was one deprecation as we moved from jQuery 2 to 3.
It looks like the `.size()` function on jQuery selectors
was replaced with the `.length` attribute.
(See http://stackoverflow.com/questions/7141334)