Recognize super old dates of birth
## Problem

When an officer entered a date of birth over 120 years ago,
a bug caused it to be ignored without showing an error.
This is a bit disconcerting for officers,
who don't have a chance to see the error.

## Solution

Accept all dates for DOB search,
even if they're unfeasably old.

Also, fix formatting of DOB error message.