Record versions of response plans
## Problem

When an officer makes an edit to a response plan,
the response plan needs to go through the approval process
before it is displayed to patrol officers.

This means that response plans are unavailable to officers
as they're awaiting review.
Officers would have no way of accessing the previously-approved
version of the response plan in the app.

## Solution

Whenever an officer makes an edit to a response plan,
create a new record in the database to store the edits,
instead of updating the existing record directly.

We can always display the most recently approved response plan,
even if there's a more updated version that's still awaiting approval.

This change also means that we are storing "versions" of response plans,
which we can later use to help Mary Perry and the legal team
track down what was visible in the app at any given time.

## Resources

https://github.com/moiristo/deep_cloneable