Remove the CSV importer, use standard DB seeds
## Problem

As we start to use the in-app form for managing response plan data,
we are no longer importing data directly from a CSV file.

The CSV file as a means for controlling our database seeds
is not very flexible.
It is difficult to update as our database structure changes,
because we need to update both a CSV file and the script that parses it.

## Solution

Change the `db/seeds.rb` script
to insert hard-coded data into the database,
instead of reading it from the CSV file.

Remove the obsolete CSV import script.