Associate Images with people, not response plans
## Problem

It is very difficult to copy over images to new response plan versions
whenever they are updated.

The libraries that we're using don't support it natively,
and if we did get it to work
we would likely see it use a lot of storage space
as it makes many copies of each image.

## Solution

* Change the `Image` model to `belong_to` a `Person`
  instead of a `ResponsePlan`.
* Accept nested attributes for images through the person model.