Revert "no officer safety" message to v0.1.4
## Problem

We want to do a slow rollout of officer safety concerns,
to allow us to add officer safety concerns to the app
before officers become dependent on having accurate information present.

## Solution

If there are no officer safety concerns for an individual,
continue displaying a message
saying that the feature is not implemented.

When we start to fill out officer safety concerns,
officers will see them gradually appear in the app.
At that point, we can switch the messaging
to more accurately reflect the feature's implementation status.

## Small Changes

Fix the background color of the officer safety section.