Add Ruby-OCI8 to pull data from the RMS Oracle db
Ruby-OCI8 is an adapter for the Oracle database connection.

Ruby-OCI8 depends on a couple of executable libraries
provided by Oracle.
To cut back on setup for the end user,
we're including these dependencies in our git repository.
We've adjusted the Dockerfile and circle.yml
to automatically install these dependencies.

We also use the same technique to host PhantomJS,
because it has been unreliable to add PhantomJS directly from BitBucket.

## Alternatives considered

Instead of storing the dependencies directly in the repository,
we also tried to run the tests on CircleCI through Docker,
and only have the dependencies installed in the Docker image.

This was infeasible, and led to a lot of hang-ups around getting Docker
installed and working correctly on CircleCI.

## Third-party dependency sources

* https://bitbucket.org/ariya/phantomjs/downloads/$PHANTOM_PACKAGE.tar.bz2
* https://github.com/sflyr/docker-sqlplus/blob/master/instantclient-basic-linux.x64-11.2.0.3.0.zip?raw=true
* https://github.com/sflyr/docker-sqlplus/blob/master/instantclient-sdk-linux.x64-11.2.0.3.0.zip?raw=true