Display people who don't have a response plan
## Feature

As a patrol officer,
When I search for an individual's information in the app
And that person doesn't have a response plan,
I want to see background information from the RMS about the person
So I can be more prepared going into the situation.

## Solution

Change the app to be more person-focused,
and less response plan-focused.
For example, we search for people instead of response plans.
We view a person's page, whether they have a response plan or not.

- Change the `/response_plans` route (back) to `/people`.
- Extract a `PeopleController` from `ResponsePlanController`,
  and copy over the `index` and `show` actions.
- If a person doesn't have a response plan,
  still display the person in the interface.