Implement searching by alias
## Problem

Often, individuals give officers fake names
that can make it difficult for the officer to identify the person.
These names are often consistent for an individual;
they'll give the same name to an officer on multiple interactions,
and they're recorded in the RMS and our system as aliases.

## Solution

When an officer searches for a name,
include search results for people
with aliases that match the search term.

See:

https://github.com/Casecommons/pg_search#searching-through-associations
https://github.com/Casecommons/pg_search#threshold