Make changing themes more consistent
## Problem

With our fancy footwork for switching themes by swapping out the `href`
attribute of our `link` tags,
some officers reported that night mode would not persist
after they left the current page.

## Solution

Revert the `href`-swapping and AJAX solution for night mode,
and go back to a regular page refresh.

Officers will likely not be swapping back and forth so often
that performance is a concern.

Reverting the javascript also restores our tests to passing,
where they had previously failed because of PhantomJS' inability to swap
stylesheets on the fly.