Add a favicon and application manifest
## Problem

The application icon in the browser and on the phone home screen
is generic, and not easily identifiable to officers.

## Solution

Create an application manifest
to display a custom icon and name for the app
when users add it to their home screen.

The manifest also tells Android to render the site
as if it were a standalone application,
which helps give the impression that it's a native app
installed on the officers' phones.

See https://developer.chrome.com/multidevice/android/installtohomescreen